Updated version to v0.1.dev2, fixed path resolve bug in compile and run and added utils.py for parac_cli
