+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Removed
+
+## [v0.1.dev2] - 2021-07-23
+
+### Added
 - Pure syntax check command (`parac syntax-check`, For info see #9 and #10)
 - Added const file `const.py` for containing constant values that are used
   throughout the module
@@ -94,5 +102,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set up a testing structure for the compiler using `pytest`
 - Created testing files for the parser and lexer
 
-[unreleased]: https://github.com/Para-C/Para-C/compare/v0.1.dev1...antlr4-dev
+[unreleased]: https://github.com/Para-C/Para-C/compare/v0.1.dev2...antlr4-dev
+[v0.1.dev2]: https://github.com/Para-C/Para-C/compare/v0.1.dev1...v0.1.dev2
 [v0.1.dev1]: https://github.com/Para-C/Para-C/releases/tag/v0.1.dev1
