+++ src/parac/__init__.py
@@ -7,7 +7,7 @@ __url__ = "https://github.com/Para-C/Para-C/"
 __author__ = "Luna Klatzer"
 __author_email__ = "luna.klatzer@gmail.com"
 __license__ = "GNU GENERAL PUBLIC LICENSE v3.0"
-__version__ = "v0.1.dev1"
+__version__ = "v0.1.dev2"
 __code_name__ = ""
 __release__ = f"{__code_name__} {__version__}"
 __copyright__ = "Luna Klatzer"
