+++ src/parac_cli/__main__.py
@@ -1,27 +1,26 @@
 # coding=utf-8
 """ Main file of the Para-Compiler"""
-import shutil
 import time
 import asyncio
-from typing import Union, Tuple
+from pathlib import Path
+from typing import Union
 import click
 import colorama
 import logging
-import os
 from os import PathLike
 from rich.progress import Progress
 
 from parac import RUNTIME_COMPILER
 from parac.exceptions import InvalidArgumentsError
-from parac.const import DEFAULT_BUILD_PATH, DEFAULT_DIST_PATH
 from parac.util import (cli_keep_open_callback, escape_ansi_args,
                         requires_init, is_c_compiler_ready,
                         cli_initialise_c_compiler, abortable)
 from parac.logging import (get_rich_console as console, print_result_banner,
                            cli_create_prompt, cli_format_default,
                            init_rich_console, print_init_banner)
-from parac.compiler import (ParacCompiler, ProgramCompilationProcess,
+from parac.compiler import (ProgramCompilationProcess,
                             BasicProcess, FinishedProcess)
+from .utils import cli_run_output_dir_validation, cli_resolve_path
 
 __all__ = [
     'cli_create_process',
@@ -38,11 +37,11 @@ colorama.init(autoreset=True)
 
 @abortable(step="Setup", reraise=True, preserve_exception=True)
 def cli_create_process(
-        file: Union[str, PathLike],
+        file: Union[str, PathLike, Path],
         encoding: str,
-        log_path: Union[str, PathLike],
-        build_path: str,
-        dist_path: str
+        log_path: Union[str, PathLike, Path],
+        build_path: Union[str, PathLike, Path],
+        dist_path: Union[str, PathLike, Path]
 ) -> ProgramCompilationProcess:
     """
     Creates a compilation process, which can be used for compiling Para-C code
@@ -52,6 +51,10 @@ def cli_create_process(
     if not RUNTIME_COMPILER.log_initialised:
         RUNTIME_COMPILER.init_logging_session(log_path)
 
+    # Resolving path and stripping whitespaces
+    file: str = cli_resolve_path(file).strip()
+    build_path: str = cli_resolve_path(build_path).strip()
+    dist_path: str = cli_resolve_path(dist_path).strip()
     return ProgramCompilationProcess(file, encoding, build_path, dist_path)
 
 
@@ -69,72 +72,6 @@ def create_basic_process(
         RUNTIME_COMPILER.init_logging_session(log_path)
 
     return BasicProcess(file, encoding)
-
-
-@abortable(step="Validating Output", reraise=True)
-def _err_dir_already_exists(folder: Union[str, PathLike]) -> bool:
-    """ Asks the user whether the build folder should be overwritten """
-    _input = console().input(
-        f"[bright_yellow] > [bright_white]The {folder} "
-        "folder already exists. Overwrite data? (y\\N): "
-    ).lower() == 'y'
-    return _input
-
-
-def _check_destination(
-        output_type: str,
-        default_path: Union[str, PathLike],
-        overwrite: bool
-) -> str:
-    """
-    Validates the destination and checks whether the specified output
-    folder is available. If the folder already exists it will show a prompt
-    to the user what should be done about the existing folder.
-
-    :returns: The path to the folder
-    """
-    output = default_path
-    if not os.path.exists(default_path):
-        os.mkdir(default_path)
-    elif len(os.listdir(default_path)) > 0:
-        # If the overwrite is set to False then a prompt will appear
-        if overwrite is False:
-            overwrite = _err_dir_already_exists(output_type)
-
-        if overwrite:
-            shutil.rmtree(default_path)
-            os.mkdir(default_path)
-        else:
-            counter = 2
-            while os.path.exists(f"{os.getcwd()}/{output_type}_{counter}"):
-                counter += 1
-            output = f"{os.getcwd()}/{output_type}_{counter}"
-            os.mkdir(output)
-    return output
-
-
-def cli_run_output_dir_validation(
-        overwrite_build: bool,
-        overwrite_dist: bool
-) -> Tuple[str, str]:
-    """ Validates whether the output folder /build/ and /dist/ can be used
-
-    :param overwrite_build: If set to True if a build folder already exists
-                            it will be deleted and overwritten
-    :param overwrite_dist: If set to True if a dist folder already exists
-                           it will be deleted and overwritten
-    """
-    build_path = _check_destination(
-        "build",
-        DEFAULT_BUILD_PATH,
-        overwrite_build
-    )
-    dist_path = _check_destination(
-        "dist",
-        DEFAULT_DIST_PATH,
-        overwrite_dist
-    )
-    return build_path, dist_path
 
 
 async def run_process(p: ProgramCompilationProcess) -> FinishedProcess:
